Refactor GroupJob -> extension on Job
